+++ tests/test_helper.rb
@@ -3,18 +3,22 @@
 require "minitest/autorun"
 require "minitest/pride"
 
-$LOAD_PATH << File.expand_path('../lib', __dir__)
+begin
+  require "pry-byebug"
+rescue LoadError
+end
+
+$LOAD_PATH << File.expand_path("../lib", __dir__)
 Thread.abort_on_exception = true
 
 require "n_plus_one_control/minitest"
 require "benchmark"
 require "active_record"
 require "factory_girl"
-require "pry-byebug"
 
-ActiveRecord::Base.establish_connection(adapter: 'sqlite3', database: ':memory:')
+ActiveRecord::Base.establish_connection(adapter: "sqlite3", database: ":memory:")
 
-Dir["#{File.dirname(__FILE__)}/../spec/support/**/*.rb"].each { |f| require f }
+Dir["#{File.dirname(__FILE__)}/../spec/support/**/*.rb"].sort.each { |f| require f }
 
 module TransactionalTests
   def setup